# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  success: boolean;
  message: string;
  /** @nullable */
  devCode?: string | null;
}

export interface RegistrationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  mobile: string;
  /** @minLength 8 */
  password: string;
}

export type OtpRequestType = typeof OtpRequestType[keyof typeof OtpRequestType];


export const OtpRequestType = {
  email: 'email',
  mobile: 'mobile',
  reset: 'reset',
} as const;

export interface OtpRequest {
  email: string;
  type: OtpRequestType;
}

export type OtpVerificationInputType = typeof OtpVerificationInputType[keyof typeof OtpVerificationInputType];


export const OtpVerificationInputType = {
  email: 'email',
  mobile: 'mobile',
  reset: 'reset',
} as const;

export interface OtpVerificationInput {
  email: string;
  type: OtpVerificationInputType;
  /** @pattern ^[0-9]{6}$ */
  code: string;
}

/**
 * @nullable
 */
export type OtpVerificationResponseNextStep = typeof OtpVerificationResponseNextStep[keyof typeof OtpVerificationResponseNextStep] | null;


export const OtpVerificationResponseNextStep = {
  mobile: 'mobile',
  authenticator: 'authenticator',
  success: 'success',
} as const;

export interface OtpVerificationResponse {
  success: boolean;
  message: string;
  /** @nullable */
  nextStep?: OtpVerificationResponseNextStep;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  mobile: string;
  emailVerified: boolean;
  mobileVerified: boolean;
  createdAt: string;
}

export interface LoginResult {
  success: boolean;
  message: string;
  user: User;
}

export interface ForgotPasswordInput {
  email: string;
}

export interface ResetPasswordInput {
  email: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
  /** @minLength 8 */
  password: string;
}