# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SecureID authentication API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Account creation, verification, and authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a new SecureID account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Account created and verification started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid registration details
        "409":
          description: An account with the email already exists
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send a verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid OTP request
        "429":
          description: OTP cooldown or rate limit
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify an email or mobile code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerificationInput"
      responses:
        "200":
          description: Code accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerificationResponse"
        "400":
          description: Invalid or expired code
        "429":
          description: Too many attempts
  /auth/resend-otp:
    post:
      operationId: resendOtp
      tags: [auth]
      summary: Resend a verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: Code resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid OTP request
        "429":
          description: OTP cooldown or rate limit
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to a verified account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          description: Invalid login details
        "401":
          description: Invalid email or password
        "403":
          description: Account still needs verification
  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the signed-in account
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Authentication required
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Revoke the current session
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Start a password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password with a reset code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired reset code
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        devCode:
          type: ["string", "null"]
      required: [success, message]
    RegistrationInput:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 80
        lastName:
          type: string
          minLength: 1
          maxLength: 80
        email:
          type: string
          format: email
        mobile:
          type: string
          minLength: 7
          maxLength: 20
        password:
          type: string
          minLength: 8
      required: [firstName, lastName, email, mobile, password]
    OtpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        type:
          type: string
          enum: [email, mobile, reset]
      required: [email, type]
    OtpVerificationInput:
      type: object
      properties:
        email:
          type: string
          format: email
        type:
          type: string
          enum: [email, mobile, reset]
        code:
          type: string
          pattern: "^[0-9]{6}$"
      required: [email, type, code]
    OtpVerificationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        nextStep:
          type: ["string", "null"]
          enum: [mobile, authenticator, success, null]
      required: [success, message]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [success, message, user]
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        mobile:
          type: string
        emailVerified:
          type: boolean
        mobileVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - mobile
        - emailVerified
        - mobileVerified
        - createdAt
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordInput:
      type: object
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: "^[0-9]{6}$"
        password:
          type: string
          minLength: 8
      required: [email, code, password]