# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface LoginResult {
  success: boolean;
  message: string;
  user: User;
}