# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResponse {
  success: boolean;
  message: string;
  /** @nullable */
  devCode?: string | null;
}