# lib/api-zod/src/generated/types/otpRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { OtpRequestType } from './otpRequestType';

export interface OtpRequest {
  email: string;
  type: OtpRequestType;
}