# lib/api-zod/src/generated/types/otpVerificationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { OtpVerificationInputType } from './otpVerificationInputType';

export interface OtpVerificationInput {
  email: string;
  type: OtpVerificationInputType;
  /** @pattern ^[0-9]{6}$ */
  code: string;
}