# lib/api-zod/src/generated/types/otpVerificationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */
import type { OtpVerificationResponseNextStep } from './otpVerificationResponseNextStep';

export interface OtpVerificationResponse {
  success: boolean;
  message: string;
  /** @nullable */
  nextStep?: OtpVerificationResponseNextStep;
}