# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  lastName: string;
  email: string;
  /**
     * @minLength 7
     * @maxLength 20
     */
  mobile: string;
  /** @minLength 8 */
  password: string;
}