# lib/api-zod/src/generated/types/resetPasswordInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordInput {
  email: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
  /** @minLength 8 */
  password: string;
}