# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * SecureID authentication API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  firstName: string;
  lastName: string;
  email: string;
  mobile: string;
  emailVerified: boolean;
  mobileVerified: boolean;
  createdAt: string;
}